+++ src/senaite/storage/permissions.py
@@ -1,12 +1,18 @@
 # -*- coding: utf-8 -*-
 
-TransitionMoveContainer = "senaite.storage: Transition: Move Container"
-
+# Permissions when context is a sample
 TransitionStoreSample = "senaite.storage: Transition: Store Sample"
 TransitionRecoverSample = "senaite.storage: Transition: Recover Sample"
+TransitionBookOutSample = "senaite.storage: Transition: Book out Sample"
 
+# Permissions when context is a storage samples container
 TransitionAddSamples = "senaite.storage: Transition: Add Samples"
 TransitionRecoverSamples = "senaite.storage: Transition: Recover Samples"
+TransitionBookOutSamples = "senaite.storage: Transition: Book out Samples"
 
+# Permissions when context is a storage content
 TransitionDeactivate = "senaite.storage: Transition: Deactivate"
 TransitionActivate = "senaite.storage: Transition: Activate"
+
+# Permissions when context is a storage container
+TransitionMoveContainer = "senaite.storage: Transition: Move Container"