+++ src/senaite/storage/workflow/sample/events.py
@@ -18,14 +18,14 @@
 # Copyright 2019-2020 by it's authors.
 # Some rights reserved, see README and LICENSE.
 
-from senaite.core.workflow import SAMPLE_WORKFLOW
-from senaite.storage import api as _api
-from senaite.storage import logger
-from zope.lifecycleevent import modified
-
 from bika.lims import api
 from bika.lims.utils import changeWorkflowState
 from bika.lims.workflow import doActionFor as do_action_for
+from senaite.core.workflow import SAMPLE_WORKFLOW
+from senaite.storage import api as _api
+from senaite.storage import logger
+from senaite.storage.api import pause_snapshots_for
+from senaite.storage.api import resume_snapshots_for
 
 
 def after_store(sample):
@@ -39,17 +39,36 @@ def after_store(sample):
     parts = primary.getDescendants()
 
     # Partitions in some statuses won't be considered
-    skip = ['cancelled', 'stored', 'retracted', 'rejected']
+    skip = ["cancelled", "stored", "retracted", "rejected"]
     parts = filter(lambda part: api.get_review_status(part) not in skip, parts)
     if not parts:
         # There are no partitions left, transition the primary
         do_action_for(primary, "store")
 
 
+def after_book_out(sample):
+    """Event triggered after "book_out" transition takes place for a given sample
+    """
+    primary = sample.getParentAnalysisRequest()
+    if not primary:
+        return
+
+    # Store primary sample if its partitions have been stored
+    parts = primary.getDescendants()
+
+    # Partitions in some statuses won't be considered
+    skip = ["cancelled", "stored", "retracted", "rejected"]
+    parts = filter(lambda part: api.get_review_status(part) not in skip, parts)
+    if not parts:
+        # There are no partitions left, transition the primary
+        do_action_for(primary, "book_out")
+
+
 def after_recover(sample):
     """Unassigns the sample from its storage container and "recover". It also
     transitions the sample to its previous state before it was stored
     """
+    # remove from container
     container = _api.get_storage_sample(api.get_uid(sample))
     if container:
         container.remove_object(sample)
@@ -58,10 +77,10 @@ def after_recover(sample):
 
     # Transition the sample to the state before it was stored
     previous_state = get_previous_state(sample) or "sample_due"
+    # Note: we pause the snapshots here because events are fired next
+    pause_snapshots_for(sample)
     changeWorkflowState(sample, SAMPLE_WORKFLOW, previous_state)
-
-    # Notify the sample has ben modified
-    modified(sample)
+    resume_snapshots_for(sample)
 
     # Reindex the sample
     sample.reindexObject()
@@ -75,14 +94,14 @@ def after_recover(sample):
     parts = primary.getDescendants()
 
     # Partitions in some statuses won't be considered.
-    skip = ['stored']
+    skip = ["stored", "booked_out"]
     parts = filter(lambda part: api.get_review_status(part) in skip, parts)
     if not parts:
         # There are no partitions left, transition the primary
         do_action_for(primary, "recover")
 
 
-def get_previous_state(instance, omit=("stored",)):
+def get_previous_state(instance, omit=("stored", "booked_out")):
     # Get the review history, most recent actions first
     history = api.get_review_history(instance)
     for item in history: