+++ src/senaite/storage/workflow/samplescontainer/events.py
@@ -23,3 +23,10 @@ def after_recover_samples(samples_container):
     """
     for sample in samples_container.get_samples():
         samples_container.remove_object(sample)
+
+
+def after_book_out_samples(samples_container):
+    """Removes all samples contained in this samples container
+    """
+    for sample in samples_container.get_samples():
+        samples_container.remove_object(sample)